Move piranha-arguments, undo piranha folder
