+++ generic/piranha/src/test/mod.rs
@@ -18,8 +18,8 @@ use log::info;
 
 use crate::config::CommandLineArguments;
 use crate::models::source_code_unit::SourceCodeUnit;
-use crate::piranha::flag_cleaner::FlagCleaner;
-use crate::piranha::piranha_arguments::PiranhaArguments;
+use crate::models::piranha_arguments::PiranhaArguments;
+use crate::piranha::FlagCleaner;
 use crate::utilities::{eq_without_whitespace, find_file, initialize_logger, read_file};
 
 pub mod test_piranha_java;
